+++ src/app/services/multipleChoice/multiple-choice.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { HttpClientService } from './http/http-client.service';
+import { HttpClientService } from '../http/http-client.service';
 import { Observable } from 'rxjs';
 
 @Injectable({
